Fix mobs not appearing with Packet underflow error
The latest HeavenMS source sends 16 null bytes, not 22 as was used in this packet handling code.  This causes wrong data when reading the packet and trying to read past the end of the packet resulting in "Packet Error: Stack underflow at 42, Opcode: 236" & "Packet Error: Stack underflow at 43, Opcode: 238".  Now mobs will show in the client.